% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevd.R
\name{fevd}
\alias{fevd}
\title{Forecast Error Variance Decomposition}
\usage{
fevd(x, ...)
}
\arguments{
\item{x}{A fitted object.}

\item{...}{Other arguments passed to methods}
}
\description{
Forecast Error Variance Decomposition
}
