% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd.R
\name{hd}
\alias{hd}
\title{Historical Decomposition}
\usage{
hd(object, ...)
}
\arguments{
\item{object}{A fitted var object.}

\item{...}{Other arguments passed to methods}
}
\description{
Historical Decomposition
}
