% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\alias{index<-}
\alias{set_index}
\title{Retrieve/Replace the index}
\usage{
index(x, ...)

index(x) <- value

set_index(x, value)
}
\arguments{
\item{x}{An object.}

\item{...}{Further arguments passed to methods.}

\item{value}{An ordered vector of the same length as the `index' attribute of x.}
}
\description{
Retrieve or replace the index of an  object.
}
