% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf}
\alias{irf}
\title{Impulse Response Functions}
\usage{
irf(x, ...)
}
\arguments{
\item{x}{A fitted object.}

\item{...}{Other arguments passed to method}
}
\description{
Impulse Response Functions
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{smgenerics:::methods_rd("irf")}
}

